idrs oough hh ,.., and code clenaup ,..
