+++ SS2-Project/Assets/Languages/en/SS2Lang_ItemsLunar_en.json
@@ -32,7 +32,7 @@
     "SS2_ITEM_RELICOFECHELON_LORE": "No no Youngling, you are here too soon! I haven't the time to prepare. Please...Come back later.",
 
     "SS2_ITEM_PRIMAL_BIRTHRIGHT_NAME": "Primal Birthright",
-    "SS2_ITEM_PRIMAL_BIRTHRIGHT_DESC": "Grants <style=cIsUtility>{0}</style> <style=cStack>(+{1} per stack)</style> Legendary Chest(s) <style=cIsUtility>on each stage</style>, but you must purchase it within <style=cDeath>{2} (+{3}s per stack) seconds of entering the stage</style> and <style=cDeath>cannot leave without claiming your birthright</style>.",
+    "SS2_ITEM_PRIMAL_BIRTHRIGHT_DESC": "Grants <style=cIsUtility>{0}</style> <style=cStack>(+{1} per stack)</style> Legendary Chest(s) <style=cIsUtility>on each stage</style>, but you must purchase it within <style=cDeath>{2} </style><style=cStack>(+{3}s per stack)</style> <style=cDeath>seconds of entering the stage</style> and <style=cDeath>cannot leave without claiming your birthright</style>.",
     "SS2_ITEM_PRIMAL_BIRTHRIGHT_PICKUP": "Grants a <style=cIsUtility>Legendary Chest</style> on each stage, <style=cDeath>but you cannot leave without claiming your birthright</style> and <style=cDeath>must claim it within a time limit</style>.",
     "SS2_ITEM_PRIMAL_BIRTHRIGHT_LORE": "The symbols on the parchment are foreign to you. You do not understand what they are, but somehow, you understand the meaning they carry. Power. Will. Authority, and a message. \n\n\"The right to rule is earned through strength and wit, and those alone. It cannot be granted by another. It cannot be granted by birth nor blood. Only strength justifies itself.\"\nYou look back down. The flickering lights of the hull cast strange shapes in the shard. Reflected shadows against its glow, coalescing into shapes that coalesce into figures.\n\nTwo shadows stumble aimlessly around in the reflection, against the backdrop of dawn. Short, stubby shadows. A seed falls into one's hair, and it laughs and plants it back into the soil. A seed falls into another's hands. It gently pries the thing open, to learn what will grow.\n\nThe shadows lengthen as the sun moves through the sky. Long, imposing shadows against twilight. Positioned perfectly, like an eerie gate to the light.\n\nOne shadow begins to grow faster still, stretching until the whole world is a blanket of darkness. The lone shadow thrashes against its prison, indistinguishable against the sea of blackness, fiercely refusing to fade away.\n\nOne shadow remains in the light. It stands in front of the now rising sun, eclipsing its light. Rays of golden sun wrap around it, forming a coronal halo for the shadow. Standing out starkly against the world around it- blinding, searing, formless. Selfish. It does not move.\n\nYou squint your eyes, and the symbols hold a new meaning.\n\n\"It isn't fair.\""
