+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Artifact/Cognation/matCognation.mat
@@ -85,7 +85,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _RemapTex:
-        m_Texture: {fileID: 2800000, guid: e29fa9a4cebd523459a97a5665356b44, type: 3}
+        m_Texture: {fileID: 2800000, guid: e4f7f64a750d83cb8a2828300d4a994a, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _SliceAlphaTex:
@@ -98,11 +98,11 @@ Material:
         m_Offset: {x: 0, y: 0}
     m_Ints: []
     m_Floats:
-    - _AlphaBias: 0.1
-    - _AlphaBoost: 1
+    - _AlphaBias: 0.2559012
+    - _AlphaBoost: 3.9
     - _BlueChannelBias: 0
     - _BlueChannelSmoothness: 0
-    - _Boost: 10
+    - _Boost: 3.2
     - _BumpScale: 1
     - _CalcTextureAlphaOn: 0
     - _CloudOffsetOn: 1
@@ -117,7 +117,7 @@ Material:
     - _DisableRemapOn: 0
     - _DistortionStrength: 0.32
     - _DitherOn: 0
-    - _DstBlend: 1
+    - _DstBlend: 10
     - _EliteBrightnessMax: 1
     - _EliteBrightnessMin: 0
     - _EliteIndex: 0
@@ -174,7 +174,7 @@ Material:
     - _SpecularStrength: 0
     - _SplatmapOn: 0
     - _SplatmapTileScale: 1
-    - _SrcBlend: 1
+    - _SrcBlend: 5
     - _UVSec: 0
     - _UseUV1On: 0
     - _VertexAlphaOn: 0
@@ -189,5 +189,5 @@ Material:
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _FlashColor: {r: 0, g: 0, b: 0, a: 1}
     - _RimColor: {r: 0.4383876, g: 0, b: 0.9779412, a: 0}
-    - _TintColor: {r: 1, g: 1, b: 1, a: 1}
+    - _TintColor: {r: 0.40392157, g: 0.69803923, b: 1, a: 1}
   m_BuildTextureStacks: []
