+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Characters/Chirr/idrsChirr.asset
@@ -3614,3 +3614,22 @@ MonoBehaviour:
         localAngles: {x: 59.19023, y: 65.71275, z: 132.8283}
         localScale: {x: 4.6, y: 4.4, z: 4.4}
         limbMask: 0
+  - keyAsset: {fileID: 11400000, guid: e6560c4541eb11a4893c230acd608387, type: 2}
+    keyAssetAddress:
+      m_AssetGUID: 
+      m_SubObjectName: 
+      m_SubObjectType: 
+    displayRuleGroup:
+      rules:
+      - ruleType: 0
+        followerPrefab: {fileID: 8349224457750773638, guid: 5e83a0c86eda7804fb0199ddad3a22e1,
+          type: 3}
+        followerPrefabAddress:
+          m_AssetGUID: 
+          m_SubObjectName: 
+          m_SubObjectType: 
+        childName: Base
+        localPos: {x: -1.24363, y: 3.3376, z: -1.40198}
+        localAngles: {x: 0, y: 307.7183, z: 0}
+        localScale: {x: 0.00363, y: 0.00363, z: 0.00363}
+        limbMask: 0
