+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Characters/Executioner2/idrsExecutioner2.asset
@@ -3146,3 +3146,22 @@ MonoBehaviour:
         localAngles: {x: 279.6506, y: 337.7838, z: 97.09126}
         localScale: {x: 0.3, y: 0.3, z: 0.3}
         limbMask: 0
+  - keyAsset: {fileID: 11400000, guid: e6560c4541eb11a4893c230acd608387, type: 2}
+    keyAssetAddress:
+      m_AssetGUID: 
+      m_SubObjectName: 
+      m_SubObjectType: 
+    displayRuleGroup:
+      rules:
+      - ruleType: 0
+        followerPrefab: {fileID: 8349224457750773638, guid: 5e83a0c86eda7804fb0199ddad3a22e1,
+          type: 3}
+        followerPrefabAddress:
+          m_AssetGUID: 
+          m_SubObjectName: 
+          m_SubObjectType: 
+        childName: Base
+        localPos: {x: -0.44483, y: 0.03239, z: -0.48138}
+        localAngles: {x: 350.2263, y: 273.4144, z: 103.3224}
+        localScale: {x: 0.00159, y: 0.00159, z: 0.00159}
+        limbMask: 0
