+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Characters/NemCommando/idrsNemCommando.asset
@@ -3392,3 +3392,22 @@ MonoBehaviour:
         localAngles: {x: 359.9454, y: 352.8523, z: 190.9242}
         localScale: {x: 0.25, y: 0.25, z: 0.25}
         limbMask: 0
+  - keyAsset: {fileID: 11400000, guid: e6560c4541eb11a4893c230acd608387, type: 2}
+    keyAssetAddress:
+      m_AssetGUID: 
+      m_SubObjectName: 
+      m_SubObjectType: 
+    displayRuleGroup:
+      rules:
+      - ruleType: 0
+        followerPrefab: {fileID: 8349224457750773638, guid: 5e83a0c86eda7804fb0199ddad3a22e1,
+          type: 3}
+        followerPrefabAddress:
+          m_AssetGUID: 
+          m_SubObjectName: 
+          m_SubObjectType: 
+        childName: Base
+        localPos: {x: -2.26868, y: -0.59192, z: -2.22017}
+        localAngles: {x: 334.7269, y: 269.1805, z: 105.214}
+        localScale: {x: 0.002, y: 0.002, z: 0.002}
+        limbMask: 0
