+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Characters/NemMercenary/idrsNemMercenary.asset
@@ -3602,3 +3602,22 @@ MonoBehaviour:
         localAngles: {x: 0, y: 0, z: 0}
         localScale: {x: 0, y: 0, z: 0}
         limbMask: 0
+  - keyAsset: {fileID: 11400000, guid: e6560c4541eb11a4893c230acd608387, type: 2}
+    keyAssetAddress:
+      m_AssetGUID: 
+      m_SubObjectName: 
+      m_SubObjectType: 
+    displayRuleGroup:
+      rules:
+      - ruleType: 0
+        followerPrefab: {fileID: 8349224457750773638, guid: 5e83a0c86eda7804fb0199ddad3a22e1,
+          type: 3}
+        followerPrefabAddress:
+          m_AssetGUID: 
+          m_SubObjectName: 
+          m_SubObjectType: 
+        childName: Base
+        localPos: {x: -0.44773, y: 1.83707, z: -0.27039}
+        localAngles: {x: 0, y: 283.3109, z: 0}
+        localScale: {x: 0.00196, y: 0.00196, z: 0.00196}
+        limbMask: 0
