+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Items/PrimalBirthright/DisplayPrimalBirthrightFollower.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aa592dbffd078727f959f9f088e65e64
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
