+++ SS2-Project/Assets/Starstorm2/AssetStorm/Base/Items/PrimalBirthright/matPrimalBirthrightMain.mat
@@ -12,7 +12,7 @@ Material:
   m_ValidKeywords: []
   m_InvalidKeywords: []
   m_LightmapFlags: 4
-  m_EnableInstancingVariants: 0
+  m_EnableInstancingVariants: 1
   m_DoubleSidedGI: 0
   m_CustomRenderQueue: -1
   stringTagMap: {}
