+++ SS2-Project/Assets/Starstorm2/Components/BirthrightObjectiveTimer.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using RoR2;
 using RoR2.UI;
 using SS2;
@@ -10,6 +11,8 @@ namespace Starstorm2.Components
 {
     public class BirthrightObjectiveTimer : NetworkBehaviour
     {
+        public static BirthrightObjectiveTimer instance;
+        
         [SyncVar] 
         public float timer;
         [SyncVar] 
@@ -19,8 +22,11 @@ namespace Starstorm2.Components
 
         private void OnEnable()
         {
+            instance = this;
+            
             if (NetworkServer.active)
             {
+                // im not sure how stacking should work here.,,. being generous to players ! each one gets their own initial stacking ! 
                 foreach (PlayerCharacterMasterController pcmc in PlayerCharacterMasterController._instancesReadOnly)
                 {
                     int? itemCount = pcmc.master?.inventory?.GetItemCountEffective(SS2Content.Items.PrimalBirthright.itemIndex);
@@ -31,19 +37,19 @@ namespace Starstorm2.Components
                 }
             }
             
+            //since the object spawns on enable itll spawn on initial pickup and dont want the storm to go off ,.,. 
             if (PrimalBirthrightObjectiveToken.instanceList.Count == 0)
             {
                 completed = true;
                 return;
             }
-            
-            PrimalBirthright.Behavior.timerComponent = this;
-            ObjectivePanelController.collectObjectiveSources += PrimalBirthright.OnCollectObjectiveSources;
+
+            ObjectivePanelController.collectObjectiveSources += OnCollectObjectiveSources;
         }
 
         private void OnDisable()
         {
-            ObjectivePanelController.collectObjectiveSources -= PrimalBirthright.OnCollectObjectiveSources;
+            ObjectivePanelController.collectObjectiveSources -= OnCollectObjectiveSources;
         }
 
         private void FixedUpdate()
@@ -84,12 +90,53 @@ namespace Starstorm2.Components
 
         public void TryRemoveObjective()
         {
-            //only remove the objective if the player hasnt activated the survive text .,,. 
             if (!spawnedMeteors)
             {
-                ObjectivePanelController.collectObjectiveSources -= PrimalBirthright.OnCollectObjectiveSources;
+                ObjectivePanelController.collectObjectiveSources -= OnCollectObjectiveSources;
                 completed = true;
             }
+        }
+        
+        public static void OnCollectObjectiveSources(CharacterMaster master, List<ObjectivePanelController.ObjectiveSourceDescriptor> objectiveSourcesList)
+        {
+            var newObjective = new ObjectivePanelController.ObjectiveSourceDescriptor
+            {
+                master = master,
+                objectiveType = typeof(PrimalBirthrightObjectiveTracker),
+                source = PrimalBirthright.primalToken
+            };
+
+            objectiveSourcesList.Add(newObjective);
+        }
+    }
+    
+    public class PrimalBirthrightObjectiveTracker : ObjectivePanelController.ObjectiveTracker
+    {
+        public override string GenerateString()
+        {
+            if (!BirthrightObjectiveTimer.instance)
+            {
+                // this *shouldnt* happen but ,.., swagever go my base case, ,. 
+                return string.Format(Language.GetString("SS2_BIRTHRIGHT_OBJECTIVE"), PrimalBirthrightObjectiveToken.instanceList.Count, PrimalBirthright.birthrightCompletionTime + (PrimalBirthright.birthrightCompletionTimeStacking * (Util.GetItemCountGlobal(SS2Content.Items.PrimalBirthright.itemIndex, false) - 1)));;
+            }
+            
+            if (BirthrightObjectiveTimer.instance.timer <= 0)
+            {
+                return string.Format(Language.GetString("SS2_BIRTHRIGHT_OBJECTIVEFAILED"), PrimalBirthrightObjectiveToken.instanceList.Count);
+            }
+            
+            string text = string.Format(Language.GetString("SS2_BIRTHRIGHT_OBJECTIVE"), PrimalBirthrightObjectiveToken.instanceList.Count, ((int)(BirthrightObjectiveTimer.instance.timer) + 1));
+            if (BirthrightObjectiveTimer.instance.timer < 30 && (int)(Time.time * 12f) % 2 == 0)
+            {
+                text = $"<style=cDeath>{text}</style>";
+            }
+
+            return text;
+        }
+
+        public override bool IsDirty()
+        {
+            return true;
         }
     }
 }
